host-vm: update start.sh messages
Update the start.sh messages to document the correct flow.

  attempt -> remote to boot and run w/out efidisk

  local -> attempt to reprogram the efidisk

The key is to ALWAYS hit ESC and reset UEFI after the NBFT has been
programmed.  This applies after booting and rebooting with the remote
option, which boots without the efidisk.

UEFI has a serious bug in that it doesn't load the boot variables on a
cold boot.  You must reset/warm boot everytime to get the variables once
they have been programmed.

Signed-off-by: John Meneghini <jmeneghi@redhat.com>