target-vm: make 'rh-start' work with netinstall ISOs
Until now, the target-vm would not boot if installed using
the 'rh-start' unattended install method. The reason for this
was that the 'livemedia-creator' emulates the boot drive as an
ordinary SATA drive, so dracut does not see a reason to include
the NVMe driver in initramfs. However, the 'vm.sh' script
emulates an NVMe drive. Thus, the system would fail to boot
due to the missing device driver and be stuck in an infinite
loop, waiting for the root filesystem.

The fix was introduced mainly by installing the
'dracut-config-generic' package, which allows to run dracut in
'non-hostonly' mode. This mode allows for including more than
the minimal required set of files in initramfs.

Signed-off-by: Michal Rábek <mrabek@redhat.com>