target-vm: get rid of repeated requests for sudo password in rh-start
The 'livemedia-creator' tool needs a considerable amount of time to
install the user-chosen OS onto the qcow2 disk image. During that time,
sudo priviledges granted by the sudo command tend to expire, resulting
in a new request for a sudo password. This is not very user-friendly.

The main problem is that the 'livemedia-creator' *must* be called as
root. This causes the produced drive to be owned by root and thus being
inaccessible to a non-root user.

This patch fixes this by first building the qcow2 disk image in the
'target-vm/.lorax' directory and then copying it to the
'target-vm/disks' directory. The added benefit of this is that it is now
possible to run 'make wipe' to wipe the 'rh-boot.qcow2' drive and then
restore it to a freshly-installed state.

Signed-off-by: Michal Rábek <mrabek@redhat.com>