bugfix: [ $? -ne 0 ] cannot be used with set -e
There were several commands being called with an if statement checking
their retun code. This is of course worng, because most scripts are now
running in bash with -e, which is going to treat every non-zero return
code as a fatal failure. Thus, that if statement is never going to be
executed.

This bug caused 'setup.sh net' to be non-functional and always exit
early on a freshly installed machine.

Signed-off-by: Michal Rábek <mrabek@redhat.com>