setup: 'setup.sh net' failing partially due to duplicated commands
The 'setup.sh' script is running with 'set -e' now. These duplicate
commands would assume that the connection linked to the default network
interface was still up, however that would not be the case. This bug can
be reproduced even in 3206ee4, but at that point it would result in a
mere error message that was easy to miss in the logs.

Signed-off-by: Michal Rábek <mrabek@redhat.com>