setup: add a quick test script
This patch builds upon all the previously introduced improvements.
It adds a single subcommand to the 'setup.sh' script that automatically
spins up and configures both virtual machines. Interractive prompts
shall guide the user through the configuration setup. The fully
automated pahse of the setup ends with the start of the host-vm. The
user must do the remaining steps manually, as there is currently no way
to automate the (fragile) UEFI boot menu setup.

Signed-off-by: Michal Rábek <mrabek@redhat.com>