target-vm: allow for observing rh-start over VNC
rh-start automatic installation was using the livemedia-creator from the
lorax suite. That program did not allow passing the '-vnc :0' argument
to the underlying QEMU process. Thus, the automated installation could
not have been observed over VNC when ran on a remote machine.

This patch solves that by (again) relying on pure QEMU to run the
installation from the ISO file. It also refactors a signifficant part of
the target-vm/vm.sh script so that it makes use of UNIX-style flag
arguments instead of confusing positional arguments.

The patch also introduces two new flags:
  '-i iso' that allows to supply any ISO file to 'vm.sh install'
  '--foreground' that makes QEMU run in foreground; blocks the terminal

Related-To: timberland-sig/rh-linux-poc#21
Signed-off-by: Michal Rábek <mrabek@redhat.com>