target-vm: fix start-nvme-target.service failing
The 'nvmetcli restore' command was intentionally removed from
'start-nvme-target.sh' due to it failing on the second and subsequent
boots of the 'target-vm'. Thus, the nvme soft target was being started
only once.

It was discovered, that the true cause of the issue was an incorrect
ordering of systemd services. If the 'start-nvme-target.service' was
attempted to be started too early, it would fail to open a network
socket and fail to start.

Setting the service to wait for the 'network-online.target' resolves the
issue.

Related-To: timberland-sig/rh-linux-poc#22
Signed-off-by: Michal Rábek <mrabek@redhat.com>